int8 types

+++ packages/hardhat/contracts/GameLogic.sol
@@ -9,10 +9,10 @@ contract GameLogic is AccessControl {
     address public postTokenAddress;
     address public voteTokenAddress;
     // upvote post address => (user address, count)
-    mapping (address => mapping (address => uint256)) public upVoteMap;
+    mapping (address => mapping (address => int8)) public upVoteMap;
     //mapping (address => uint256) public upVoteMap;
     // downvote post address => count
-    mapping (address => mapping (address => uint256)) public downVoteMap;
+    mapping (address => mapping (address => int8)) public downVoteMap;
     // user address => nft address
     mapping (address => address) public postAddress;
 
@@ -29,9 +29,8 @@ contract GameLogic is AccessControl {
     }
 
     function vote(address _postAddress, int8 votes) public {
-        uint256 currentUpVoteCount = upVoteMap[_postAddress][msg.sender];
-        uint256 currentDownVoteCount = downVoteMap[_postAddress][msg.sender];
-
+        int8 currentUpVoteCount = upVoteMap[_postAddress][msg.sender];
+        int8 currentDownVoteCount = downVoteMap[_postAddress][msg.sender];
         if (votes < 0) {
             downVoteMap[_postAddress][msg.sender] = currentDownVoteCount + uint256(-votes);
         } else if (votes > 0) {
